+++ mrmustard/lab_dev/wires.py
@@ -12,28 +12,22 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-""" Classes for supporting tensor network functionalities."""
+"""`Wires` class for handling the connectivity of an object in a circuit."""
 
 from __future__ import annotations
 
 from typing import Iterable, Optional
+
 import numpy as np
+
 from mrmustard import settings
 
 # pylint: disable=protected-access
-
-__all__ = ["Wire", "Wires"]
-
-Wire = int
-r"""
-An integer representing a wire in a tensor network.
-TODO: Leaving this because it's used in other files for type-hinting,
-    but let's decide whether we want this (and ``Mode``) or not.
-"""
+# pylint: disable=import-outside-toplevel
 
 
 class Wires:
-    r"""A class with wire functionality for tensor network applications.
+    r"""`Wires` class for handling the connectivity of an object in a circuit.
     In MrMustard, ``CircuitComponent``s have a ``Wires`` object as attribute
     to handle the wires of the component and to connect components together.
 
@@ -43,27 +37,42 @@ class Wires:
     input bra modes --->|   circuit   |---> output bra modes
     input ket modes --->|  component  |---> output ket modes
                         ---------------
+    Each of the four groups can be empty. In particular, the wires of a state have
+    no inputs and the wires of a measurement have no outputs. Similarly,
+    a standard unitary transformation has no bra wires.
 
-    The "standard order" mentioned below is output_bra for all modes,
-    input_bra for all modes, output_ket for all modes, input_ket for all modes.
-    We use this order when we inline the ids into a list, or when we reproduce the
-    init args from the ids.
+    We refer to these four groups in a "standard order":
+        0. output_bra for all modes
+        1. input_bra for all modes
+        2. output_ket for all modes
+        3. input_ket for all modes
 
     A ``Wires`` object can return subsets (views) of itself. Available subsets are:
-
     - input/output  (wires on input/output side)
     - bra/ket       (wires on bra/ket side)
     - modes         (wires on the given modes)
-    - id_subset     (wires with the given ids)
 
     For example, ``wires.input`` returns a ``Wires`` object with only the input wires
-    (on bra and ket sides and on all the modes). Or ``wires.input.bra[(1,2)] returns a
+    (on bra and ket sides and on all the modes).
+    Note that these can be combined together: ``wires.input.bra[(1,2)] returns a
     ``Wires`` object with only the input bra wires on modes 1 and 2.
-    Note these are views of the original ``Wires`` object, i.e. we can set the ``ids``
-    on the views and it will be set on the original, e.g. ``wires1.output.ids = wires2.input.ids``.
+    Note these are views of the original ``Wires`` object, i.e. we can set the ``ids`` on the
+    views and it will be set on the original, e.g. this is a valid way to connect two sets
+    of wires by setting their ids to be equal: ``wires1.output.ids = wires2.input.ids``.
 
-    ``Wires`` can also be added to one another, which returns a new ``Wires`` object with
-    the wires of both objects combined (if there are duplicates, an error is raised).
+    A very useful feature of the ``Wires`` class is the support for the right shift
+    operator ``>>``. This allows us to connect two ``Wires`` objects together as
+    ``wires1 >> wires2``. This will return the ``Wires`` object of the two
+    wires objects connected together as if they were in a circuit:
+               ____________                             ____________
+    in bra --->|  wires1  |---> out bra ---> in bra --->|  wires2  |---> out bra
+    in ket --->|          |---> out ket ---> in ket --->|          |---> out ket
+               ------------                             ------------
+    The returned ``Wires`` object will contain the surviving wires of the two
+    ``Wires`` objects and it will raise an error if there are overlaps between
+    the surviving wires. This is especially useful for handling the ordering of the
+    wires when connecting components together: we are always guaranteed that a
+    ``Wires`` object will provide the wires in the standard order.
 
     Args:
         modes_out_bra (Iterable[int]): The output modes on the bra side.
@@ -87,22 +96,19 @@ class Wires:
         modes_out_ket = modes_out_ket or []
         modes_in_ket = modes_in_ket or []
 
-        self._modes = list(
+        self._modes = sorted(
             set(modes_out_bra) | set(modes_in_bra) | set(modes_out_ket) | set(modes_in_ket)
         )
         randint = settings.rng.integers  # MM random number generator
-        ob = {m: randint(1, 2**62) if m in modes_out_bra else 0 for m in self._modes}
-        ib = {m: randint(1, 2**62) if m in modes_in_bra else 0 for m in self._modes}
-        ok = {m: randint(1, 2**62) if m in modes_out_ket else 0 for m in self._modes}
-        ik = {m: randint(1, 2**62) if m in modes_in_ket else 0 for m in self._modes}
-        if len(self._modes) > 0:
-            self._id_array = np.array([[ob[m], ib[m], ok[m], ik[m]] for m in self._modes])
-        else:
-            self._id_array = np.zeros((0, 4), dtype=np.int64)
+        outbra = {m: randint(1, 2**62) if m in modes_out_bra else 0 for m in self._modes}
+        inbra = {m: randint(1, 2**62) if m in modes_in_bra else 0 for m in self._modes}
+        outket = {m: randint(1, 2**62) if m in modes_out_ket else 0 for m in self._modes}
+        inket = {m: randint(1, 2**62) if m in modes_in_ket else 0 for m in self._modes}
+        self._id_array = np.array([[outbra[m], inbra[m], outket[m], inket[m]] for m in self._modes])
         self._mask = np.ones_like(self._id_array)  # multiplicative mask
 
     def _args(self):
-        r"returns the same args one needs to initialize this object."
+        r"""Returns the same args one needs to initialize this object."""
         ob_modes = np.array(self._modes)[self._id_array[:, 0] > 0].tolist()
         ib_modes = np.array(self._modes)[self._id_array[:, 1] > 0].tolist()
         ok_modes = np.array(self._modes)[self._id_array[:, 2] > 0].tolist()
@@ -110,7 +116,7 @@ class Wires:
         return tuple(ob_modes), tuple(ib_modes), tuple(ok_modes), tuple(ik_modes)
 
     @classmethod
-    def _from_data(cls, id_array, modes, mask=None):
+    def _from_data(cls, id_array: np.ndarray, modes: list[int], mask=None):
         r"""Private class method to initialize Wires object from the given data."""
         w = cls()
         w._id_array = id_array
@@ -149,7 +155,7 @@ class Wires:
 
     @property
     def modes(self) -> list[int]:
-        "The set of modes of the populated wires."
+        "The set of modes spanned by the populated wires."
         return [m for m in self._modes if any(self.id_array[self._modes.index(m)] > 0)]
 
     @property
@@ -163,22 +169,22 @@ class Wires:
 
     @property
     def input(self) -> Wires:
-        "A view of this Wires object without output wires"
+        "A view of self without output wires"
         return self._view(masked_cols=(0, 2))
 
     @property
     def output(self) -> Wires:
-        "A view of this Wires object without input wires"
+        "A view of self without input wires"
         return self._view(masked_cols=(1, 3))
 
     @property
     def ket(self) -> Wires:
-        "A view of this Wires object without bra wires"
+        "A view of self without bra wires"
         return self._view(masked_cols=(0, 1))
 
     @property
     def bra(self) -> Wires:
-        "A view of this Wires object without ket wires"
+        "A view of self without ket wires"
         return self._view(masked_cols=(2, 3))
 
     @property
@@ -186,18 +192,14 @@ class Wires:
         r"""
         The adjoint (ket <-> bra) of this wires object.
         """
-        if len(self.modes) > 0:
-            return self._from_data(self._id_array[:, [2, 3, 0, 1]], self._modes, self._mask)
-        return self
+        return self._from_data(self._id_array[:, [2, 3, 0, 1]], self._modes, self._mask)
 
     @property
     def dual(self) -> Wires:
         r"""
         The dual (in <-> out) of this wires object.
         """
-        if len(self.modes) > 0:
-            return self._from_data(self._id_array[:, [1, 0, 3, 2]], self._modes, self._mask)
-        return self
+        return self._from_data(self._id_array[:, [1, 0, 3, 2]], self._modes, self._mask)
 
     def copy(self) -> Wires:
         r"""A copy of this Wires object with new ids."""
@@ -219,7 +221,7 @@ class Wires:
         return self._from_data(combined_array, sorted(modes_rows), np.ones_like(combined_array))
 
     def __bool__(self) -> bool:
-        return True if len(self.ids) > 0 else False
+        return len(self.ids) > 0
 
     def __eq__(self, other: Wires) -> bool:
         return np.array_equal(self.id_array > 0, other.id_array > 0)
@@ -233,44 +235,46 @@ class Wires:
     def __lshift__(self, other: Wires) -> Wires:
         return (other.dual >> self.dual).dual  # how cool is this
 
+    @staticmethod
+    def _outin(si, so, oi, oo):
+        r"""Returns the output and input wires of the composite object made by connecting
+        two single-mode (ket or bra) objects like --|self|-- and --|other|--
+        At this stage we are guaranteed that the configurations `|self|--  |other|--`  and
+        `--|self|  --|other|` (which would be invalid) have already been excluded.
+        """
+        if bool(so) == bool(oi):  # if the inner wires are either both there or both not there
+            return np.array([oo, si], dtype=np.int64)
+        elif not si and not so:  # no wires on self
+            return np.array([oo, oi], dtype=np.int64)
+        else:  # no wires on other
+            return np.array([so, si], dtype=np.int64)
+
     def __rshift__(self, other: Wires) -> Wires:
-        r"""Returns a new Wires object with the wires of self and other combined as two
-        components in a circuit: the output of self connects to the input of other wherever
-        they match. All surviving wires are arranged in the standard order.
-        A ValueError is raised if there are any surviving wires that overlap."""
         all_modes = sorted(set(self.modes) | set(other.modes))
         new_id_array = np.zeros((len(all_modes), 4), dtype=np.int64)
-        for i, m in enumerate(all_modes):
-            if m in self.modes and m in other.modes:
-                # m-th row of self and other (self output bra = sob, etc...)
-                sob, sib, sok, sik = self._mode(m)
-                oob, oib, ook, oik = other._mode(m)
-                errors = {
-                    "output bra": sob and oob and not oib,
-                    "output ket": sok and ook and not oik,
-                    "input bra": oib and sib and not sob,
-                    "input ket": oik and sik and not sok,
-                }
-                if any(errors.values()):
-                    position = [k for k, v in errors.items() if v][0]
-                    raise ValueError(f"wire overlap at {position} of mode {m}")
-                if bool(sob) == bool(oib):  # if the inner wires are both there or both not there
-                    new_id_array[i] += np.array([oob, sib, 0, 0])
-                elif not sib and not sob:
-                    new_id_array[i] += np.array([oob, oib, 0, 0])
-                else:
-                    new_id_array[i] += np.array([sob, sib, 0, 0])
-                if bool(sok) == bool(oik):
-                    new_id_array[i] += np.array([0, 0, ook, sik])
-                elif not sik and not sok:
-                    new_id_array[i] += np.array([0, 0, ook, oik])
-                else:
-                    new_id_array[i] += np.array([0, 0, sok, sik])
-            elif m in self.modes and not m in other.modes:
-                new_id_array[i] += self._mode(m)
-            elif m in other.modes and not m in self.modes:
-                new_id_array[i] += other._mode(m)
-        return self._from_data(np.abs(new_id_array), all_modes)
+
+        for m in set(self.modes) & set(other.modes):
+            sob, sib, sok, sik = self._mode(m)  # m-th row of self
+            oob, oib, ook, oik = other._mode(m)  # m-th row of other
+
+            out_bra_issue = sob and oob and not oib
+            out_ket_issue = sok and ook and not oik
+            if out_bra_issue or out_ket_issue:
+                raise ValueError(f"Output wire overlap at mode {m}")
+            in_bra_issue = oib and sib and not sob
+            in_ket_issue = oik and sik and not sok
+            if in_bra_issue or in_ket_issue:
+                raise ValueError(f"Input wire overlap at mode {m}")
+
+            new_id_array[all_modes.index(m)] = np.hstack(
+                [self._outin(sib, sob, oib, oob), self._outin(sik, sok, oik, ook)]
+            )
+        for m in set(self.modes) - set(other.modes):
+            new_id_array[all_modes.index(m)] = self._mode(m)
+        for m in set(other.modes) - set(self.modes):
+            new_id_array[all_modes.index(m)] = other._mode(m)
+
+        return self._from_data(new_id_array, all_modes)
 
     def __repr__(self) -> str:
         ob_modes, ib_modes, ok_modes, ik_modes = self._args()
@@ -327,10 +331,13 @@ class Wires:
         html += "</table>"
 
         try:
-            from IPython.display import display, HTML  # pylint: disable=import-outside-toplevel
+            from IPython.core.display import (
+                HTML,
+                display,
+            )
 
             display(HTML(html))
         except ImportError as e:
             raise ImportError(
-                "To display the wires in a jupyter notebook you need to `pip install IPython`"
+                "To display the wires in a jupyter notebook you need to `pip install IPython` first."
             ) from e